// src/app/api-services-v2/model/orderStatusHistoryResponseDTO.ts
/**
 * Bestellsystem API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OrderStatus } from './orderStatus';


export interface OrderStatusHistoryResponseDTO { 
    status?: OrderStatus;
    timestamp?: string;
}
export namespace OrderStatusHistoryResponseDTO {
}